+++ c++/pod5_format_pybind/repack/repack_functions.h
@@ -82,6 +82,8 @@ arrow::Result<ReadReadData> read_read_data(
             columns.num_reads_since_mux_change->Value(batch_row);
         auto const & time_since_mux_change = columns.time_since_mux_change->Value(batch_row);
         auto const & open_pore_level = columns.open_pore_level->Value(batch_row);
+        auto const & expected_open_pore_level = columns.expected_open_pore_level->Value(batch_row);
+        auto const & selected_read_level = columns.selected_read_level->Value(batch_row);
         auto const & num_samples = columns.num_samples->Value(batch_row);
 
         auto const & pore_type_index = source_reads_pore_type_column.Value(batch_row);
@@ -117,7 +119,9 @@ arrow::Result<ReadReadData> read_read_data(
             predicted_scaling_shift,
             num_reads_since_mux_change,
             time_since_mux_change,
-            open_pore_level);
+            open_pore_level,
+            expected_open_pore_level,
+            selected_read_level);
         result.signal_durations.emplace_back(num_samples);
 
         auto const signal_rows = std::static_pointer_cast<arrow::UInt64Array>(