+++ CHANGELOG.md
@@ -18,6 +18,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Possibility recovery misinterpreting uninitialised buffer as useful data.
 
+### Added
+
+- Added new field `expected_open_pore_level`, containing the expected open pore level for this read.
+- Added new field `selected_read_level`, containing the selected pore level for this read.
+
 ## [0.3.43]
 
 ### Changed